+++ source/source_io/CMakeLists.txt
@@ -49,6 +49,7 @@ list(APPEND objects_advanced
     module_unk/berryphase.cpp
     module_wannier/to_wannier90.cpp
     module_wannier/to_wannier90_pw.cpp
+    module_wannier/to_wannier90_pw_io.cpp
     module_wannier/to_w90_lcao_pw.cpp
     module_wannier/to_wannier90_lcao.cpp
     module_wannier/to_wannier90_lcao_io.cpp
